requirements.txt
alembic==1.8.0
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
redis==4.3.4
SQLAlchemy==1.4.39
Werkzeug==2.1.2
wrapt==1.14.1
